Setting Tree now holds state
Value is now stored in settings tree
settings tree is now responsible for calling callbacks when value 
changes from inside the ui, or out.
accomplishes this by using 2 accessors. one is added to the tree, and 
updates the ui when changed. the other is passed to the ui, and may 
optionaly run callbacks when the value is changed.